Topic08_Final_Paul_GenerativeModellingWithInverseHeatDissipation/requirements.txt
blobfile==2.0.1
mpi4py==3.1.5
numpy==1.24.2
opencv_python==4.7.0.68
Pillow==10.2.0
torch==1.13.1
torchvision==0.14.1